sintax changes, files renamed to be smaller
